maps/ctmap: moved CtKey{4,6} to types.go
